requirements.txt
streamlit==1.32.0
paho-mqtt==1.6.1
requests==2.31.0
Flask==2.3.2
python-dotenv==1.0.1
gunicorn==21.2.0